test(fermiedge): scope Qt widget keepalive
